+++ BaseApi.Business/Repositories/RecipeIngredientRepository.cs
@@ -23,8 +23,11 @@ namespace Pasteleria.Business.Repositories
         public async Task DeleteAsync(Guid id)
         {
             var result = await GetByIdAsync(id);
-            _context.RecipeIngredients.Remove(result);
-            await _context.SaveChangesAsync();
+            if (result != null)
+            {
+                _context.RecipeIngredients.Remove(result);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task<List<RecipeIngredient>> GetAllAsync()
@@ -32,16 +35,14 @@ namespace Pasteleria.Business.Repositories
             return await _context.RecipeIngredients.AsNoTracking().ToListAsync();
         }
 
-        public async Task<RecipeIngredient> GetByIdAsync(Guid id)
+        public async Task<RecipeIngredient?> GetByIdAsync(Guid id)
         {
-            var result = await _context.RecipeIngredients.FirstOrDefaultAsync(ri => ri.Id == id);
-            return result == null ? throw new KeyNotFoundException($"RecipeIngredient with ID {id} not found.") : result;
+            return await _context.RecipeIngredients.FirstOrDefaultAsync(ri => ri.Id == id);
         }
 
         public async Task UpdateAsync(RecipeIngredient dto)
         {
-            var existingRecipeIngredient = await GetByIdAsync(dto.Id);
-            _context.Entry(existingRecipeIngredient).CurrentValues.SetValues(dto);
+            _context.RecipeIngredients.Update(dto);
             await _context.SaveChangesAsync();
         }
     }
